Move classes and structs into their own dedicated files
