+++ Assets/Chainblocks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 474479ab8f6583048b47a3bbd012a7a7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
