+++ Assets/Chainblocks/CBCore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d3eab72c83016e4493fb86d26c4e9c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
