+++ Assets/Chainblocks/CBCoreExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec68d5053886ced4fabef4ed4ecc15fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
