+++ Assets/Chainblocks/CBType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de30ae308199c44468e3d29a4ca5c2a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
