+++ Assets/Chainblocks/CBVar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9dd715d3f4fcfe042a870f8b42f2ef60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
