+++ Assets/Chainblocks/CBlock.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 257f13157d8dd4d43828297a31000870
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
