+++ Assets/Chainblocks/Delegates.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03f89b9837984c6499e012aeb06264ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
