+++ Assets/Chainblocks/ExternalVariable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3f0e78851b03a70478ac51e186591b0f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
