+++ Assets/Chainblocks/Native.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e547454c47527447aa908cd1cac63261
+guid: b1723d8972e178b45bff8841258c7b66
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
