+++ Assets/Chainblocks/Variable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 185bf365cb04c3448bed90a0b162fd94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
